refactor: move offset.hpp to Offset.hpp
- Move the file src/Editor/offset/offset.hpp to src/Offset/Offset.hpp
