+++ src/CMakeLists.txt
@@ -21,7 +21,7 @@ target_sources(kilo
         "${PROJECT_SOURCE_DIR}/src/Editor/config/config.hpp"
         "${PROJECT_SOURCE_DIR}/src/Editor/config/config.cpp"
 
-        "${PROJECT_SOURCE_DIR}/src/Editor/offset/offset.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Offset/Offset.hpp"
 
         "${PROJECT_SOURCE_DIR}/src/Terminal/Terminal.hpp"
         "${PROJECT_SOURCE_DIR}/src/Terminal/Terminal.cpp"
