+++ tests/CMakeLists.txt
@@ -31,7 +31,7 @@ target_sources(tests
         "${PROJECT_SOURCE_DIR}/src/Editor/config/config.hpp"
         "${PROJECT_SOURCE_DIR}/src/Editor/config/config.cpp"
 
-        "${PROJECT_SOURCE_DIR}/src/Editor/offset/offset.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Offset/Offset.hpp"
         Offset.test.cpp
 
         "${PROJECT_SOURCE_DIR}/src/Terminal/Terminal.hpp"
