+++ tests/Offset.test.cpp
@@ -1,4 +1,4 @@
-#include "Editor/offset/offset.hpp"
+#include "Offset/Offset.hpp"
 
 #include <gtest/gtest.h>
